% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_clock.R
\name{clock_secs}
\alias{clock_secs}
\title{Convert "MM:SS" to seconds remaining}
\usage{
clock_secs(x)
}
\arguments{
\item{x}{Character vector.}
}
\description{
Convert "MM:SS" to seconds remaining
}
\keyword{internal}
