% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conferences.R
\name{conference_members}
\alias{conference_members}
\title{Members of one conference in one season}
\usage{
conference_members(season, code)
}
\arguments{
\item{season}{Season year.}

\item{code}{Conference code (e.g. "CC").}
}
\value{
Output of \code{\link[=parse_conference_standings]{parse_conference_standings()}}.
}
\description{
Members of one conference in one season
}
