% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conferences.R
\name{conference_name_from_title}
\alias{conference_name_from_title}
\title{Conference name from a standings page title}
\usage{
conference_name_from_title(html, season)
}
\arguments{
\item{html}{Parsed standings page.}

\item{season}{Season year.}
}
\description{
"2025 Centennial Conference football standings" -> "Centennial
Conference"; some titles omit "football" ("2025 Conference of New
England standings").
}
\keyword{internal}
