% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_plays.R
\name{default_user_agent}
\alias{default_user_agent}
\title{Default browser User-Agent}
\usage{
default_user_agent()
}
\value{
A character scalar.
}
\description{
Default browser User-Agent
}
\keyword{internal}
