% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_plays.R
\name{fetch_refused}
\alias{fetch_refused}
\title{Has d3football refused a request since the last \code{\link[=reset_fetch_state]{reset_fetch_state()}}?}
\usage{
fetch_refused()
}
\description{
Has d3football refused a request since the last \code{\link[=reset_fetch_state]{reset_fetch_state()}}?
}
\keyword{internal}
