% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pbp.R
\name{is_line_score}
\alias{is_line_score}
\title{Is this table the line score?}
\usage{
is_line_score(t)
}
\arguments{
\item{t}{A data frame.}
}
\value{
TRUE / FALSE.
}
\description{
A table with at least 2 rows and either a "Final" column or a first
column headed "Scoring" (a game stopped early has no "Final" column).
}
\keyword{internal}
