% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pbp.R
\docType{data}
\name{kept_row_types}
\alias{kept_row_types}
\title{Event types kept in the play-by-play table}
\format{
An object of class \code{character} of length 5.
}
\usage{
kept_row_types
}
\description{
Event types kept in the play-by-play table
}
\keyword{internal}
