% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_plays.R
\name{n_requests}
\alias{n_requests}
\title{Number of network requests made since the last \code{\link[=reset_fetch_state]{reset_fetch_state()}}}
\usage{
n_requests()
}
\description{
Number of network requests made since the last \code{\link[=reset_fetch_state]{reset_fetch_state()}}
}
\keyword{internal}
