% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{other_team}
\alias{other_team}
\title{The other team}
\usage{
other_team(x, teams)
}
\arguments{
\item{x}{Character vector of canonical team names.}

\item{teams}{The game's two canonical team names.}
}
\description{
The other team
}
\keyword{internal}
