% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pbp.R
\docType{data}
\name{pbp_columns}
\alias{pbp_columns}
\title{Output column order}
\format{
An object of class \code{character} of length 67.
}
\usage{
pbp_columns
}
\description{
cfbfastR-aligned. See \code{docs/schema.md} for the data dictionary.
}
\keyword{internal}
