% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_plays.R
\name{reset_fetch_state}
\alias{reset_fetch_state}
\title{Reset the request counter and the "refused" flag (start of a run)}
\usage{
reset_fetch_state()
}
\description{
Reset the request counter and the "refused" flag (start of a run)
}
