% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_plays.R
\name{tables_on}
\alias{tables_on}
\title{Convert every HTML table on a page to a list of data frames}
\usage{
tables_on(html)
}
\arguments{
\item{html}{A parsed HTML document (from \code{\link[=fetch_html]{fetch_html()}}).}
}
\value{
A list of data frames, one per \verb{<table>} on the page.
}
\description{
Convert every HTML table on a page to a list of data frames
}
\keyword{internal}
