% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_outcomes.R
\docType{data}
\name{text_yard_pattern}
\alias{text_yard_pattern}
\title{Yardline pattern in play text}
\format{
An object of class \code{character} of length 1.
}
\usage{
text_yard_pattern
}
\description{
"to the UCHI45", "at RMCFB32", "to the McD28", "to the Wolves05": a token
that starts with a capital letter, then the yard number. Tokens can be
upper-case codes, mixed case ("McD"), or nicknames ("Wolves").
}
\keyword{internal}
