% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_plays.R
\name{throttle}
\alias{throttle}
\title{Wait so network requests are at least \code{d3ballR.delay} seconds apart}
\usage{
throttle()
}
\description{
Wait so network requests are at least \code{d3ballR.delay} seconds apart
}
\keyword{internal}
