% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_outcomes.R
\name{token_regex}
\alias{token_regex}
\title{Regex alternation matching a game's play-text team tokens}
\usage{
token_regex(tokens)
}
\arguments{
\item{tokens}{Token strings (names of \code{\link[=infer_text_team]{infer_text_team()}}'s output).}
}
\description{
Regex alternation matching a game's play-text team tokens
}
\keyword{internal}
